refactor metrics rendet/time
